Fix circular import, assert orb counts per level, fix a few naming errors.
